re-organize db server
